+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   * 增强：`DateFieldConvertor` 支持 `Instant` 类型的转换：https://gitee.com/troyzhxu/bean-searcher/pulls/9
   * 新增：`EnumParamConvertor` 支持 `String/Emun` 类型的参数自动转换为枚举序号
   * 优化：`PreparedStatement.setObject(..)` 方法报错时，仍然打印 SQL 日志
+* Bean Searcher Boot Starter
+  * 自动配置 `EnumParamConvertor`
+* Bean Searcher Solon Plugin
+  * 自动配置 `EnumParamConvertor`
 
 ### 🐛 Bug Fixes